+++ src/mainwindow.h
@@ -937,4 +937,6 @@ private:
             bool multiNoteMenuEntriesOnly = false);
 
     void updateCurrentNoteTextHash();
+
+    void centerAndResize();
 };